3863 selectAnnotationsForPath
